pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strix-gateway"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "cryptography>=42.0.0",
]

[project.scripts]
strix = "strix_gateway.cli.main:cli"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=7.0.0",
    "greenlet>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["strix_gateway"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::pytest.PytestUnhandledThreadExceptionWarning"]

[tool.coverage.run]
source = ["strix_gateway"]
branch = true
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
show_missing = true
